make installation more automatic

+++ install.sh
@@ -1,4 +1,14 @@
-#!/bin/sh
+#!/bin/bash
+
+# install dependencies
+pip install pycrypto
+
+# check if "$HOME/.local/bin" is in the "$PATH" if no then add it to the "$PATH"
+if [ "$(echo "$PATH" | grep -c "\./local/bin")" -lt 1 ]; then
+  [ -d "$HOME/.local/bin" ] || mkdir -p "$HOME/.local/bin" # create "$HOME/.local/bin" if doesn't exist
+  echo "PATH=\"$HOME/.local/bin:$PATH\"" >> "$HOME/.bashrc"
+  [ -f "$HOME/.bashrc" ] && source "$HOME/.bashrc"
+fi
 
 chmod +x jenc.py
-ln -s "$(pwd)/jenc.py" "$HOME/.local/bin/jenc"
+[ -f "$HOME/.local/bin/jenc" ] || ln -s "$(pwd)/jenc.py" "$HOME/.local/bin/jenc"
